requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
crewai==0.22.5
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
kaggle==1.5.16
huggingface-hub==0.19.4
PyGithub==2.1.1
tavily-python==0.3.3
openai==1.6.1
pydantic==2.5.2
typing-extensions==4.9.0
markdown==3.5.1